Backend/requirements.txt
Flask==2.2.5
flask-cors==3.0.10
tensorflow==2.14.0
numpy
Pillow
pymongo==4.6.0
python-dotenv==1.0.0
PyJWT==2.8.1
bcrypt==4.1.2
requests==2.31.0
